+++ lib/wechat/shake_around.rb
@@ -14,6 +14,12 @@ require 'wechat/shake_around/beacon_page_relation'
 require 'wechat/shake_around/page_report'
 require 'wechat/shake_around/page_daily_report'
 
+require 'wechat/shake_around/group'
+require 'wechat/shake_around/device_group_relation'
+
+require 'wechat/shake_around/device_report'
+require 'wechat/shake_around/device_daily_report'
+
 module Wechat
   module ShakeAround
   end